+++ va108xx/examples/embassy/src/bin/async-spi.rs
@@ -6,11 +6,11 @@ use embassy_time::{Duration, Ticker};
 use once_cell::sync::OnceCell;
 
 use va108xx_hal::{
-    gpio::{Output, PinState},
+    gpio,
     pac::{self, interrupt},
-    pins::{PinsA, PinsB},
+    pins,
     prelude::*,
-    spi,
+    spi, InterruptConfig,
 };
 
 const SYSCLK_FREQ: Hertz = Hertz::from_raw(50_000_000);
@@ -29,11 +29,11 @@ async fn main(_spawner: Spawner) {
     // Safety: Only called once here.
     va108xx_hal::embassy_time::init(dp.tim23, dp.tim22, SYSCLK_FREQ);
 
-    let porta = PinsA::new(dp.porta);
-    let _portb = PinsB::new(dp.portb);
-    let mut led0 = Output::new(porta.pa10, PinState::Low);
-    let mut led1 = Output::new(porta.pa7, PinState::Low);
-    let mut led2 = Output::new(porta.pa6, PinState::Low);
+    let porta = pins::PinsA::new(dp.porta);
+    let _portb = pins::PinsB::new(dp.portb);
+    let mut led0 = gpio::Output::new(porta.pa10, gpio::PinState::Low);
+    let mut led1 = gpio::Output::new(porta.pa7, gpio::PinState::Low);
+    let mut led2 = gpio::Output::new(porta.pa6, gpio::PinState::Low);
 
     let spi_clk_cfg = spi::ClockConfig::from_clk(50.MHz(), 1.MHz()).unwrap();
     let mut spi_cfg = spi::Config::default();
@@ -41,11 +41,7 @@ async fn main(_spawner: Spawner) {
     let (sck, mosi, miso) = (porta.pa31, porta.pa30, porta.pa29);
     let spi = spi::Spi::<u8>::new_for_spi0(dp.spia, (sck, miso, mosi), spi_cfg);
 
-    let mut spi = spi.into_async(Some(va108xx_hal::InterruptConfig::new(
-        va108xx_hal::pac::Interrupt::OC2,
-        true,
-        true,
-    )));
+    let mut spi = spi.into_async(Some(InterruptConfig::new(pac::Interrupt::OC2, true, true)));
     SPI_TOKEN.set(spi.bank_id()).unwrap();
     let mut ticker = Ticker::every(Duration::from_secs(1));
     let buf: [u8; 4] = [0xAA; 4];